+++ src/toisto/command/__init__.py
@@ -1,0 +1,4 @@
+"""Commands."""
+
+from .practice import practice
+from .show_progress import show_progress